add ability to clone empty repo
